Isolated network operations (http requests) into a object class for easier testing and organization
